Renamed modules folder to features
