+++ src/app/features/dashboard/dashboard-routing.module.ts
@@ -1,0 +1,18 @@
+import { NgModule } from '@angular/core';
+import { Routes, RouterModule } from '@angular/router';
+import { DashboardComponent } from './dashboard/dashboard.component';
+
+const dashboardRoutes: Routes = [{
+	path: '',
+	component: DashboardComponent
+}];
+
+@NgModule({
+	imports: [
+		RouterModule.forChild(dashboardRoutes)
+	],
+	exports: [
+		RouterModule
+	]
+})
+export class DashboardRoutingModule { }
